add edit pages
