+++ src/App.js
@@ -2,17 +2,19 @@ import './App.css';
 
 import {HomePage} from './pages/HomePage/HomePage'
 import {SingersPage} from './pages/SingersPage/SingersPage'
-import SingerForm from './pages/SingersPage/SingerForm'
-import TrackForm from './pages/TracksPage/TrackForm'
+import SingerForm from './pages/SingersPage/SingerEdit'
+import SingerEdit from './pages/SingersPage/SingerEdit'
+import TrackForm from './pages/TracksPage/TrackEdit'
+import TrackEdit from './pages/TracksPage/TrackEdit'
 import {TracksPage} from './pages/TracksPage/TracksPage'
 import SingerDetail from './pages/SingersPage/DetailPage'
 import TrackDetail from './pages/TracksPage/TrackDetail'
 import TranslationsPage from './pages/TranslationsPage/TranslationsPage'
 import TranslationDetail from './pages/TranslationsPage/TranslationDetail'
-import TranslationForm from "./pages/TranslationsPage/TranslationForm";
+import TranslationForm from "./pages/TranslationsPage/TranslationEdit";
 import {AuthenticationPage} from "./pages/RegistrationPage/AuthenticationPage";
 import {RegistrationPage} from "./pages/RegistrationPage/RegistrationPage";
-
+import TranslationEdit from "./pages/TranslationsPage/TranslationEdit";
 
 import {BrowserRouter as Router, Switch, Route} from 'react-router-dom';
 
@@ -39,6 +41,15 @@ function App() {
      </Switch>
      <Switch>
         <Route path="/addSinger" component={SingerForm} />
+     </Switch>
+     <Switch>
+        <Route path="/editSinger" component={SingerEdit} />
+     </Switch>
+     <Switch>
+        <Route path="/editTrack" component={TrackEdit} />
+     </Switch>
+     <Switch>
+        <Route path="/editTranslation" component={TranslationEdit} />
      </Switch>
      <Switch>
         <Route path="/addTranslation" component={TranslationForm} />
