+++ src/pages/SingersPage/DetailPage.jsx
@@ -2,6 +2,8 @@ import React from 'react';
 
 import {Footer} from '../../components/Footer'
 import {Header} from '../../components/Header'
+import {Link} from "react-router-dom";
+import {FaPencilAlt} from "react-icons/fa";
 
 
 
@@ -12,6 +14,10 @@ export class SingerDetail extends React.Component{
             <div>
                 <Header/>
                 <div className="min-vh-100">
+                    <div className="d-flex justify-content-end">
+                        <Link to={{pathname:"/editSinger",
+                                   state:this.props.location.state}}><FaPencilAlt/></Link>
+                    </div>
                     <div className="jumbotron jumbotron-fluid">
                         <div className="container">
                             <h1 className="display-4">{this.props.location.state.name}</h1>
