+++ src/pages/TranslationsPage/TranslationDetail.jsx
@@ -2,6 +2,8 @@ import React from 'react';
 
 import {Footer} from '../../components/Footer'
 import {Header} from '../../components/Header'
+import {FaPencilAlt} from "react-icons/fa";
+import {Link} from "react-router-dom";
 
 export class TranslationDetail extends React.Component{
 
@@ -10,6 +12,8 @@ export class TranslationDetail extends React.Component{
             <div>
                 <Header/>
                 <div className="jumbotron min-vh-100 text-center">
+                    <Link to={{pathname:"/editTranslation",
+                                   state:this.props.location.state}}><FaPencilAlt/></Link>
                     <p className="lead">{this.props.location.state.language}</p>
                     <hr className="my-4"/>
                     <p>{this.props.location.state.text}</p>
@@ -19,5 +23,5 @@ export class TranslationDetail extends React.Component{
         );
     }
 }
-    
+
 export default TranslationDetail
